moving the column size input inside a component
